+++ src/work/future/block/mod.rs
@@ -1,0 +1,7 @@
+// devela::work::future::block
+//
+//! Defines the private [`future_block`] standalone function.
+//
+
+mod block;
+pub(crate) use block::*;